Change the semver range resolution to be explicit about the module requiring the dependency.
Apparently webpack’s dependency module metadata contains an issuer attribute, but the issuer attribute only seems to indicate *one* of the modules requiring the dependency, which may not be the issuer under consideration. Because our semver range resolution algorithm depends on the issuer, we must pass the issuer in explicitly.